now when rearranging verses, footnote anchor updates
